colli escape-mode: fix calculation of position of potential danger
Up to now it was done by checking potential collision with 45° steps to decide
if they are left/right/top/bottom of robot. We now use the information from
ColliShape on the angles to the corners of the robot-shape to get correct
information about the position of collisions, no matter how we modify the
robot-shape (extension).
This should theoretically also work for robots with circle-shaped bases,
or robots where the laser is placed closer to the edge than to the center
of the robot.