fvutils: add vertical flipping support to JPEG compressors
Vertical flipping is necessary for example for cameras mounted upside
down on the ceiling. The proper way would be to implement this as a
filter. But especially in our current use case (application on the
Raspberry Pi) where we need to avoid processing steps and memory
transfer an extra filter would be problematic. Since we fill the buffer
for the compresion library line-wise anyway, it is realy easy to
implement the flipping that this stage by just having a different
addressing scheme.

Therefore we add this extra functionality to the JPEG compressors.