webview-ptzcam: pan/tilt/zoom camera controller webview page
The webview-ptzcam allows viewing and controlling a pan/tilt/zoom
camera like the Sony EviD100P via webview. It automatically enables
the camera on requests and handles pan/tilt/zoom commands and setting
of some effects. It relies on the webview plugin to provide the image
streams. It parks and disables the camera after some specified time
of inactiviy.