webview: explicit deep image copying and conversion
Before this change, we relied on the input camera to be JPEG compatible,
i.e. in YUV422_PLANAR colorspace. We also relied on the shared memory
camera/buffer's deep copying capability.

This commit changes this to explicitly deep copy and/or convert the
image. This has two benefits:
- we support more tpyes of input images
- we potentially convert less images, i.e. if the framerate for an image
  stream is smaller than the capture rate, we convert less images (and
  obviously even more so for single images)

On the RasPi this reduces the load for streams from 55% to about 40% CPU
load for a single stream. If no stream is active even down to 20% (the
image is still copied into the shared memory buffer within fvretriever).
This means less packet dropping for the isochronous USB transfer.