webview: support Apache Portable Runtime for password verification
This change immediately brings in support for SHA1, MD5, and bcrypt
password hash functions. It also saves us some pitfalls. Additionally to
not magically detect clear text passwords, but rather require them to be
given with a "!cleartext!" prefix which should make it more obvious that
this is a bad idea.