eclipse-clp: add path search utility
Readylog files are sometimes distributed over different
folders and different gits, therefore a utility to
search for a file in different paths is needed.
This is only a first stage, which works only for C++.

The Problem with eclipse is that external predicates
are loaded after the module is loaded. However, I
would need this predicate to locate other modules
at load time.