netcomm: recognize services after name collision resolution
If a service is registered with a name that already exists it is
automatically modified to resolve the conflict by adding a unique
number (possibly multiple times). The infrastructure already maintained
both names, but the Avahi implementation modified the original name,
making it impossible to recognize a service with a modified name later.