gazsim-timesource: allow conversion of gazebo offset times
Gazebo uses an offset time relative to the start time of the simulation.
When receiving data with such timestamps, it must be converted to be
compatible to Fawkes times.

The current method is rather crude and can cause "jumpy" times if the
real-time factor is not stable. But this applies to all times
established with the current method.