gazebo-laser: compatibility with Gazebo's own Hokuyo laser
Instead of maintaining our own model we can more comfortably rely on the
Gazebo folks to maintain the laser driver for us. Frederik and I came to
this conclusion when investigating problems that Gazebo was mixing up
the models when it unexpectedly used its own model instead of ours.
Since there method even provides more information it is beneficial to
use it.