colli: disable odd "time_factor" for drive_realization
Apparently it should be used to flatten out deviations of the
colli frequency. As the timestamp is only updated on "Drive(..)"
calls, the first of such calculations will always be wrong
(too high, which makes the colli accelerate at its limit).
Besides, it does not seem to be safe to simply multiply the
translation by the deviation-factor. We disable this for now,
maybe there will be a meaningful use for it in the future.