urdf: add "alias" links for reference frames to match tf-names in Fawkes
We now have "katana/base_frame" and "katana/tool_frame" instead of
"katana_base_frame" and "katana_gripper_tool_frame". This is now
consistent with all the other frame names throughout Fawkes.

These frames are put on top of existing ones, so they basically serve
as aliases. This is done in order to peserve compatibility with
3rd-party code that is used to the URDF model from ROS.