laser-filter: write through/select timestamp on output
The output interfaces always got a new timestamp when written. This
causes problems because the data is potentially already one or more
frames old (e.g. if synced). Therefore either write through the
timestamps (if there was a direct buffer mapping) or select an
appropriate one (in the case of min merge, three strategies have been
implemented, latest, first, or a specific one).