colli: changed transform of laser data
the laser_data of the colli where transformed in base_link instead of base_laser
and they where transformed with Time(0,0), now they are transformed at the
given time in the interface

Now the colli transformes new laser data into the odom frame
at the (in the interface) given time.

When the occ-grid is build (obstacles inserted) the laser data (new and old)
are getting transformed (with Time(0,0)) back into base_laser to be filled in the grid.
Therefore the colli doesn't need the x-,y-,ori-diff paramethers to decived how far
the robot moved since the last loop.

Further the class Laser does not exists anymore. This class didn't give much more
features as the Laser360Interface. This resultes that the colli doesn't
publishes the laser-data for ros anymore.

Conflicts (resolved):
	src/plugins/colli/colli_thread.cpp
	src/plugins/colli/search/og_laser.cpp
	src/plugins/colli/search/og_laser.h