laser: configurable time offset or scan time factor for URG
For the URG sensor one can now define a scan time factor and/or a fixed
offset for the data. This can be used if the initial acquisition time
(i.e. when the first laser beam was taken within a rotation) is desired.
We also have some offset which we could not yet explain but where
experiments have shown that a factor of 1.5 is beneficial.