navgraph: allow shortcut to target node
Include the target node in the search for a shortcut. Before there could
be the situation that the last but one node was shadowed by an obstacle.
While we could easily shortcut to the final node (and hence reach it
quickly) we wouldn't. Now we take the final node into consideration.