utils: fix file notifications and try harder to merge events
We always tried to match the regex even if there was no string at all
(length zero). This caused problems in particular with file watches.

Additionally create an event list and allow for some time (1ms) for more
events to come in to merge them to as few invocations as possible.