amcl: use laser data timestamp, buffer if impossible
We claimed that we needed to use the latest timestamp because it was
otherwise not possible to get a transform if the laser data was written
in the sensor frame. Part of the problem was that the laser plugin wrote
timestamps not when they should have been (when the data was actually
acquired). This has been fixed now, so we actually can rely on this
timestamp.

But in case there is a problem we buffer the data for one loop and
re-try (if the fresh data still won't work).