core: set thread name if supported
On Linux the non-portable function pthread_setname_np() can be used to
set a thread's name. This name is then displayed e.g. in gdb for the
"info threads" command.

On supported systems we now call this function to get more descriptive
names in gdb. Note, however, that only at most 15 characters of the name
can be displayed.

Oh and yes, this is some more awesomeness again that we wanted for years.