colli: replace ellipse/ellipse_map with general obstacle/obstacle_map
We also introduced a base-class ColliFastObstacle as parent of
ColliFastEllipse and a new ColliFastRectangle. Other modules simply
load a ColliObstacleMap and define whether it consists of rectangles
or ellipses.
Currently ColliFastRectangle constructor is empty, and the default
type of obstacles is "ellipse". So running the colli will currently
not change anything.