colli: add obstacle based on base_link and roboshape
The colli builds the grid around 1 cell which represents
base_link. The only really precise way for an obstacle size is
therefore the robo-shape itself, where we take a 180° rotated shape
and match it's "base_link" with the laser-point.

Currently, I set to use that size also for the old readings, which
makes sense in my opinion.