pcl-db: support database parameters and flags for operations
The database name is now passed explicitly. We need to distinguish
between database and collection for GridFS. The database can still be
configured in the config, but this serves as the default only.
Overriding allows for using it for accessing mongodb-log data or to
use pcl-db more explicit for storing (later) and retrieval.