openrave: fix rotation of objects by using corect quaternion fields
openrave stores quaternions in this order: w,x,y,z
So e.g. quat[0] is "w" and quat[2] is "y" and so forth. It is also
possible to call quat.x, as quat is a 4-field long vector, but that
refers to the vector-value x which is quat[0], and not to the
x-value of the quaternion (which would be the 2nd field in the vector).
Indeed, very confusing and obviously error-prone.