skills: new "align_at_table" skill, for "good" grasp positions
This skill aligns the robot to a detected table, so that the robot
looks straight at the table and keeps a defined distance (adjustable
with "dist" parameter, 0.2m is default).
This is useful to have similar "conditions" when a grasping task is
desired. Fix distances to the table can come in very handy.