skills: add align_laserht to align to laser line
We use the laserht plugin to detect the dominant line of laser reading
in front of us (by using a circle sector filtered laser range as input).
Then this little skill will align the robot to be perpendicular to the
line facing towards it. This will hopefully allow us to come up with a
solution to properly approach the dish washer.

This was joint work with Bahram.