core: don't flag a thread as bad if the flag is already set
Flagging an already flagged thread as bad and adding it to the list of
bad threads leads to multiple recoveries in the next
loop, if a thread from the previous loop still hasn't recovered.
To avoid this, check if a thread is already flagged as bad and
if so, don't check it again.