katana: initialize "theta_error" with 0 to avoid infinite IK checks
theta_error is set with every LinearGotoMessage, but not for ParkMessage.
That is usually not a problem, but in case ParkMessage is the 1st one to
arrive, and for some reason transforms are incorrect such that "park"
would fail, the undefined theta_error would result in infinite IK checks
with different theta-values.