eclipse-clp: add IndiGolog checker util
This util can be used to check the definition of actions
and fluents. It will make a basic plausibility check (e.g.
it will warn if an action was defined, but no execute
predicate for that action exists).

Unfortunately, in the current versions modifications to the
agent have to be performed to use it. These modifications prevent
correct execution of the agent. Therefore it is only possible to
either check the code or to use the agent in it's intended way.