skill ppgoto: only send StopMessage on reset if ppnavi not final
The explicit sending of the StopMessage should stop the pathplan.
Pathplan also send StopMessage to the navigator in that case. But
doing so on skill-reset means, that the skill can be FINAL and other
skills might be executed, but then we still send a StopMessage which
(due to threading) is sent to the Navigator at a later point in time.
That will override the GotoMessages that were possibly sent meanwhile.

Example: ppgoto is FINAL, agent calls relgoto. Then ppgoto sends a stop
though, which arrives after the GotoMessage from relgoto -> relgoto
FAILES.