laser-lines: detect lines in laser pointclouds
This plugin takes a point cloud (which is assumed to have been generated
from a laser sensor, i.e. all values lie in the x-y-plane, and uses
RANSAC to find lines in this data. Up to a configurable number of lines
(best fit first) are written to the blackboard.

For each line the plugin publishes a point on the line, a direction
vector, and the bearing to the point on line. Additionally it publishes
a point cloud with colored and optimized lines, and if compiled with
visual debugging also the point on line and direction vectors, and IDs.

The lines are ordered by their bearing angle (to have stable IDs when
stationary) and the point on line is optimized to be the point with the
shortest distance from the reference frame origin (typically the laser
sensor's center).