amcl: fix laser range config values
The angle range for which laser data should be fed to the particle
filter can be constrained in the config. The processing of these values
were wrong in the case that the min value was larger than the max value,
i.e. if we wanted to have a constrained "forward" angle range.