colli: allow adjusting GuaranteeTransStop from config
The calculations are not precise, which can lead to bocking, or
driving further than the actual target. The reason is not clear yet,
although it is probably due to delays and the differences between
sent motor-command and the expected behaviour after a period of time.
That is also different for different platforms or motors.

Therefore we add the possiblity to adjust the calculations via config,
until we find a proper solution for this problem.