ros-skiller: provide a skiller interface to ROS
This plugin implements a ROS action server which provides a skiller
interface to ROS. ROS clients can send arbitrary skill commands to
fawkes by sending requests to the /skiller action server. All necessary
ROS messages are defined in the ROS package fawkes_skiller_msgs. In each
loop, this plugin sends the latest request to the skiller if the request
hasn't been executed yet. If multiple commands are requested, all but
the last one are ignored. This plugin reads the skiller status and
provides a feedback to the ROS client.