imu: CruizCore data catch up and fixes
The CruizCore XG1010 driver now checks if there is more data after
reading a packet and if so reads it to catch up. This avoids latency due
to buffer congestion and increases the chance of reading fresh data.

Along the way did a few fixes and removed some debug output.