fvutils: allow zero-sized ROIs
The contains check was arguably wrong, i.e. it was off by one. If width
or height were zero, the contains check effectively checked for the
value one by using "less or equal" instead of just a "less" check.

As a corollary the extend function must add an offset of one.