colli: send one final STOP when stopping and motor values are all zero
When stopping, "STOP" is sent until the motor values can be considered
as "stopped". To make sure that these are not flucuations in the
motor-controller or anything the like, we send one final STOP to make
sure we really stop.
Especially if such a situation occurs the moment we start stopping, the
colli would think "already stopped, so do not send further STOP messages",
while the is actually still running. We saw this behaviour a few times
in athome, resuling in endless spinning of the robot.