navgraph: make constrained search more dynamic
The NavGraphNodeConstraint is now a pure virtual class which states the
node constraint interface. A static list node constraint has been added
to mimic the old behavior of maintaining a static list of blocked nodes.
An additional compute() method is called on the constraints before
searching for a path. That allows for pre-computing/cache necessary data
for following calls to check if a node is blocked.

Along the way restructure a bit, use some C++11 goodness and add missing
documentation.