ros-pcl: run in WORLDSTATE rather than in SENSOR_PROCESS hook
Running in SENSOR_PROCESS would mean that most of the time the thread
would send the data from the last loop. Even worse, it could easily
happen that it would take half-baked data from the (shared) point
clouds as most threads will be processing it in this very step.

Moving the pcl publishing thread to the WORLDSTATE hook fixes this
problem as no processing is to be expected there. Additionally we always
send the most recent data close to when it was produced.