laser-cluster: support min length of lines to filter
Lines can now be removed based on their actual length. This allows for
tolerant line parameters but still removing unwanted outliers. An
intrinsic property to laser data is that it gets more sparse the further
away the bounce back. This leads to a difficult parameterization of
lines and clusters. The line length feature should allow to remedy this
problem short of implementing a distance-biased clustering.