navgraph: basic replanning support
If constraints have been modified since the last replan interval compute
the constraints. If any change occurs, try to re-plan. If the plan is as
good or better as the current one in terms of cost of the path, use the
new plan. Otherwise, or if a planning error occurs, keep following the
old plan.