navgraph: add config flag to not abort on error
Aborting on error, e.g. if there is no navigator, is normally the right
thing. However, for cases like (off-line) debugging this might erase the
very state you need to debug. Hence add this flag, which should only be
set to false if you really know what you are doing.