navgraph: add support for edge cost constraints
Edge cost constraints can be used to increase the actual past costs of
edges. This allows to more precisely model costs on obstructions, e.g.
obstacles or robots, without sacrificying the admissibility of the used
heuristic (assuming it was admissible before). The new constraints are
queried for a cost factor, which is accepted if it is greater or equal
to 1.0.