core: fix dtor race condition on InterruptibleBarrier
An interruptible barrier could cause a problem when a thread was taking
too long, such that an interruptible barrier timed out. If now the
thread recovered and the barrier was destructed at the same time it
would cause a segfault. This in particular occured when running in
valgrind which slowed down the overall system.

Thanks to Frederik Zwilling for spotting this.