navgraph: support configurable cost factor for re-planning
Currently we only accept a new plan if it is at most as expensive as the
old plan. The effect of this is that if constraints are added at
run-time the robot will almost never take a detour around an obstacle as
it will always be more expensive (with the exception being obstacles
visible at initial planning).

Now we have a configurable factor by which the old costs are multiplied.
If then the new costs are at most the scaled old cost we accept the new
plan. This allows for at least some control over whether and when to
accept new plans.