kinova: use only libkindrv for arm control
The plugin depends on and works with libkindrv now. All the stuff
(and more) went to that lib, which started with the kinova_api files
from this plugin.

"types.h" is reduced to the types that are only used by fawkes, all the
kinova_api relevant stuff was deleted, since it is now provided by
libkindrv.