kinova: Split openrave_thread into two, for single- and dual-arm
JacoOpenraveThread is renamed to KinovaOpenraveBaseThread, and is
now a base thread, providing basic methods and members.
Depending on the setup (a config flag), we either load
a KinovaOpenraveSingleThread or a KinovaOpenraveDualThread.

KinovaOpenraveSingleThread keeps the current functionality as it was
in previous JacoOpenraveThread.

KinovaOpenraveDualThread currently loads the robot and IK-solvers for
each arm properly, and just updates the openrave-model.
Movement-planning is not yet implemented, or at least not tested!