skill jaco: support interfaces for left and right arm
A parameter "arm=" can define if this command is specified for left
or right arm, which is needed in a dual-arm setup.
Note for future improvements: It would be better to get the config entry
that specifies what kind of setup we have, and assing the interfaces
just once. That would also allow for preconditional checking for
existing interface-writers etc.

Additionally, one can add "no_wait=true" to send the skill to FINAL
without waiting for arm to be final. This is useful for uncoordinated
dual-arm control, otherwise the two arms would have to act alternatingly.