kinova: add abstract KinovaArm class, allows having other controllers/drivers
Every command that should go to an arm is passed to a concrecte derivated
class of KinovaArm. For now we have one class, namely KinovaArmKindrv, which
communicates to a real arm via libkindrv. This has two benefits:
1) The connection point to libkinrv is only at 1 place, so other parts of the
 code do not need to bother with it
2) It is interchangable! Means, we can (and will) add for example a
 KinovaArmOpenrave class that communicates with the simulated arm in openrave,
 which allows to run the kinova plugin in a "simulation mode" without real arms.

This is similar to the "controller" class of the katana plugin, although the
name "controller" is not really suitable.