kinova: revert to old final-checking, cannot check if arm is moving
The idea was to check if the arm was moving and by that decide on
"final" or not. The check was done via "get_ang_vel()", but that
returns the joint angles and not the current velocities (is under
investigation). For now return to simply checking "target coordinates
vs. current coordinates".
We might combine both checks in the future, but for now leave it this
way