kinova: make openrave_threads continuous and assing one to each arm
The openrave_threads are responsable for all the OpenRAVE related stuff.
This includes loading and updating the model to represent the current
arm configuration, using OpenRAVE's IK solver and planning.
As we want to support uncoordinated bimanual manipulation, each arm needs
its own independent thread for planning, which is why now each arm (whether
for single-arm or dual-arm setup) is assigned its own openrave_thread.
The KinovaOpenraveSingleThread class is used for this.

Additionally we add a separate thread for planning coordinated bimanual
manipulation. This is done by the KinovaOpenraveDualThread now. It is
only responsible for planning of this kind, and loading the dual-arm robot
into the openrave-environment. Usage of this feature is not implemented yet.
This is why most of the code of this class has been stripped (not supported,
or it moved to the separate KinovaOpenraveSingleThread for each arm).