kinova: openrave_thread: add FIFO queue for targets and trajectories
The target_queue is (not yet) being processed FIFO, running path-planner
for each target. If successful, the resultet trajectory is added to the
trajec_queue (FIFO) which is accessed by the goto_thread to check if a
new trajectory is to be processed.

Adding a target to target_queue automatically checks for IK solvability
(which is the boolean return value). The target is only enqueued if solvable,
and if the parameter "plan" is not explicitly set to false. Skipping planning
means here that the IK-solved target is added as a single trajectory-point
to the trajec_queue. This can be dangerous, especially for collision avoidance,
but can also be handy if the user knows what he is doing (e.g. very short
distances with the Kinova arm can be done without planning, the internal
controllers handle those straight movements pretty well).