kinova: add planning for single arm
Each openrave_thread starts planning as soon as there is a target
in the target_queue. Upon success, the trajectory is added to the
trajec_queue, which is repeatedly checked by the goto_thread, in
order to execute the trajectory.

Proper trajectory execution needs to be added though. The planning
is probably not safe, because currently it all happens in the same
openrave environment. We need clone-environments for each planner,
especially when using multiple arms.