kinova: use RefPtr for target and trajectory queues
Using RefPtr makes the pointer management easier, because we do not need
to worry about it anymore! The target/traje mutexes and queues are
moved to from openrave_thread to jaco_arm_t to make them accessible for
all other threads (required lateron for planning<->movement coordination),
which is much easier to handle with refptrs.

Added a few tyepdefs to replace all the nested list<vector<vector....>>>
for trajectory-points, trajectories and queues. Looks cleaner now.