kinova: remove trajec_queue, keep one main target_queue
The typedefs and structs have changed a little. The idea is that we have
one target_queue, where we enqueue a target (jaco_target_t). These are
processed by the goto_threads, depending of the target-type.
Meanwhile, the openrave-threads check the queue for targets that require
path-planning. In such case, they plan the path and set the trajectory for
that target in the same jaco_target_t struct. When goto_threads process
such a target, they will wait until a trajectory is calculated, and then
execute.

The queue is designed such that the trajectories that are planned ahead
consider the previous target as their starting point, so that they can
be seamlessly executed. However, we need environment cloning in openrave
(in develeopment, but seems that it is not so trivial).