core: add CircularBuffer Container
The CircularBuffer is a simple circular buffer which uses a std::deque
to store the data. It has a fixed size which it automatically maintains
by removing elements from the front of the buffer if the buffer grows
too large. It does not allow any element manipulation other than
push_back() and pop_front().