openrave: add copy constructor for OpenRaveRobot
Copies name, offsets, and OpenRaveManipulator objects.
PlannerParameters and BaseManipulation modules are created for this robot.
We also need to pass a pointer to the new OpenRaveEnvironment as a param
to the ctor, so that we access the same robot and the same manipulator in
that environment as we did with the to-be-copied object.