kinova: Do not implicitly add finger-values to target_queue
This is only relevant when finger-values are not passed along. Then we want
to use the current finger-values. Adding those to the target_queue would
use finger-values at that timestamp, which might not necessarily be the same
anymore when that target is processed lateron (it's a queue, the fingers
might have move before this target is processed)