kinova: use separate environments for planning
The default OpenRAVE-environment, which can also be displayed by the viewer,
is only used for showing the current arm configurations, movements,
and other objects that are perceived by the robot.
We use separate environments for planning, so that both can run concurrently.
This is not only useful for showing the current movement while alerady planning
the next one (would have to lock the environment for planning). We actually
need this for concurrent planning for uncoordinated bimanual manipulation.