eclipse_externals: return message ID to eclipse in bb_send_message
bb_send_message/3 sends a message to a blackboard interface. So far, we
couldn't get the ID of the message. Add the predicate bb_send_message/4,
where the fourth argument is unified with the message ID of the sent
message. This is useful in cases where we want to check the status of a
message in eclipse. As an example, when sending messages to the skiller,
we want to be able to check whether the skiller has started to process
the message.