openprs: add aspect and communication proxies
We now provide aspects to create and access OpenPRS kernels from within
Fawkes. Additionally, we have implemented proxies for message passer and
OpenPRS server communication for monitoring and injecting commands. The
latter is required e.g. for remote consultation, though we yet need to
find or create a way to read a command's output. The proxy solution is
necessary because otherwise we could not inject commands if oprs-server
was running independently.