buildsys: allow overriding of undefined symbols behavior for libs
By default all missing symbols are considered an error on linking, also
for libraries and plugins. This is useful in almost all cases. But there
are exceptions, for example when integrating with 3rd-party code.
Therefore, we at least provide a way to override the behavior by setting
the ALLOW_UNDEF_libname variable to 1.