openprs-agent: OP invocation coalescing meta OP
This meta Meta OP to coalesce marked OP instances of the same OP.
Sometimes it is desirable to activate only a single OP even if
multiple facts for which the OP triggers have been added, e.g. if
the OP itself performs steps of reduction of the data.

This is achieved by this meta OP. To use it, the OPs which should
be coalesced must have the property COALESCE set to T.