kinova: plot trajectories in viewer_env with adjustable colors
The viewer_env is the environment we see, so this is where we
want to display the planned trajectories that are about to be
executed.
The trajectory color is "red" by default. For dual-arm setups,
we plot the trajectory of the right arm in blue. These colors
can be modified, but currently only within act_thread.cpp