openrave: add method to clone objects from another environment
This clones the objects (KinBody) form one environment to another,
maintaining their states. KinBodys that are parts of a Robot are
ignored. KinBodys that do not exist in the reference environment
are deleted, such that we have a "mirrored" environment.

Implemented this method because OpenRAVE's Clone method takes too
long, clones too much or deletes/resets more than we want. This method
almost finishes instantly, while the full OpenRAVE Clone takes 4 secs.