openrave: allow adjusting filter-options for IK check
By default we check against environment-collisions. This is not
always usefule, e.g. when we just want the IK check while planning
is done at a later step (that's when we want to check for environment
collisions).
This IK filter-option can now be passed as an additional argument
to the "set_target_" methods