kinova: add error-codes and trajectory-states
Error codes were added to JacoInterface. We need this to react to
certain failures (IK fail, planning fail, ...), especiall on skill
side.

The trajectory-states are added for better handling of the target
queue with a trajectory-type target. For example, the planning-thread
can now set the state to "PLANNING_ERROR", allowing other threads to
react on it. For now, the reaction on such PLANNING_ERROR is "abort
the queue", but it could e.g. also be "retry planning" or something else.

The check for "final" now includes checking if the target-queue has
been fully processed. This makes sense, and it is also makes "final"
more robust.