openrave: Jaco model: each finger is a separate join now
Previously 2 fingers mimiced the 3rd finger, which was not very
precise to begin with. The fingeres of the real arm can be controlled
individually. This is now also possible with the OpenRAVE model.

The robot.xml file has been adapted, properly separating joints used
for IK (excludes gripper joints) and joints used for grasping (i.e.
the gripper joints).