jaco: look further back in target_queue for start of new plan
Instead of just taking the previous target (if available) as a
starting point for our new target, we look back until we find one that
contains joint-values. For example the sequence
"move1 -> open_gripper -> move2" could plan ahead for move2 with starting
point of move1's final point. If we only took the last target ("open_gripper"),
we would see that the values are invalid and use current values for planning,
which would result in a wrong pre-calculated path for move1.