openprs: support passing a GDB delay parameter to kernels
To debug OpenPRS modules it is helpful to connect to a kernel via GDB.
This can be tricky if the problem occurs early after startup. The new
argument can be passed that will instruct mod_utils to wait for a few
seconds and print an appropriate message that contains the required gdb
command.