libconfig: fix issues regarding list values
Lists of length 1 were not properly submitted due to a wrong condition
in the marshaller of the message. Retrieving specific list values for
anything other than strings did not work. Finally, determining the type
of an empty list did not work (as we fix the type on the first value).
Change the semantics so that an empty list is of any type, rather than
of no type.