jaco: always set active manipulator before planning
We restore the robot state from the viewer-environment, so some things
about the robot may change in the planner-environment, for instance
which manipulator is active. This affects both IK search (non-critical)
and planning (critical). Therefore re-set active manipulator in planning,
after the objects have been cloned and the robot state was restored.