openprs: mod_skiller - online check for exclusive control
When loaded straight from the command line, the exclusive control
acquisition message cannot be processed by the skiller for the time we
check and would therefore always fail.

Now we check when executing a skill and also try a (non-stealing)
re-acquire of the exclusive control if we don't have it (yet).