openrave: improve locking + allow setting manipulator for object-attach
The OpenRAVE environment is now locked for every action. There may be
some redundant locks, but better be safe than getting errors (which
can result in segfaults).

It is now also possible to pass the manipulator-name to AttachObject.
This is required for multi-arm setups, otherwise the object would
just be attached to the currently active manipulator.