eclipse-clp: add the possibility to simulate action execution
If simulate_execution is set, execute/2 is overwritten for all actions
in order to simulate action execution. Actions are executed by logging
the action to be executed. This also prevents the real execution of any
action (thus you do not need to define execute/2 for your action).

For sensing actions, sensing results can be provided beforehand. If no
sensing result is defined, you will be asked for the result when the
sensing action is executed.

If ask_for_exogenous_events is set, you will be asked for exogenous
events after each action execution.