robotino: support OpenRobotino API 2 (and thus Robotino 3)
This is actually a simple change which adds (additional) support for
OpenRobotino API 2. The code has been re-structured to fit the flow of
both API versions better. Calls for API 2 to gather the same data and
execute the same commands have been added. This looks much worse than it
actually is!