robotino-joystick: adapt to deadman switch feature
Do not only send commands if the joystick has been moved but all the
time. This is required to avoid the deadman switch added to the robotino
plugin in the past. Additionally we now require any button to be pressed
to actually move (safety precaution). Alternatively we monitor axis 5,
which is the RT button on our XBox 360 controller.