openrave: disable manual initialization of planner
Might fail (on in our case "does fail") due to incorrect planner
parameters, which are set by OpenRaveRobot::set_target_ methods.
Need to investigate this in the future, however, we only included
the initialization to help tracking down the error in case planning
failed. The BaseManipulation module that does the planning for us
has an own initialization routin, so we are not "missing" anything.