jaco: add method for trajectory execution to JacoArm class
We need to consecutively send all trajectory points to libkindrv, so that
they are enqueued in the arm's FIFO queue. We can't use goto_joints(..) for
that, because that acquires api-access and angular-control first. Doing that
for each trajectory-point somehow breaks the continuity in the FIFO queue.

It is also more convenient to send the whole trajectory to the class that
actually commands the arm, so the goto_thread doesn't need to bother with
that anymore.