jaco: properly enqueue targets for coordinated manipulation
Targets have an additional boolean flag "coord", which tells if the
target needs coordination with targets of other arms. This is how the
plannener and movement threads (i.e. openrave_threads and goto_threads)
tell which target they can process. If "coord==true", only the "bimanual"
threads process these targets, and for example "goto_thread" waits meanwhile.

Also fixed a bug in openrave_thread. When searching for a previous target in
queue that can be used for a starting point of planning, we need to abort if
we come upon a target with unknown configuration (e.g. TARGET_CARTESIAN that
skips planning, or a TARGET_HOME).