jaco: added path planning for coordinated manipulation
As soon as the first targets in the target queues require trajectories
and have the "coord" flat set to true, they are processed. First checked
for closest solvable IK such that there is no collision with
_solve_multi_ik(). Then plan the path, and store the trajectories if
the plan was successful.

Currently, after successful planning, we set "coord" flag to false. This
is just for debugging, as there is no goto_thread for coordinated
manipulation yet. With this, the goto-threads for uncoordinated manipulation
start executing the trajectory.

Sorting the IK solutions for shortest distance turned out to be very helpful,
basically eliminating all those weird movements. Need to add this to the
default openrave_thread as well!