jaco: plan target with joint values, not Transform
We use the transform (or to be precies, Euler coordinates) to check for
IK solution. The openrave plugin remembers this and would plan with these,
i.e. call BaseManipulations's MoveToHandPosition. That method does not sort
IK solutions by distance though, so we modify our target such that the
openrave plugin uses MoveActiveJoints command.