openrave: fix configuration comparison when looking for best IK
OpenRAVE already provides a method to calculate the difference of
joint configuration. That one is more precice, e.g. considers
"circular" joints and stuff. We use that method now.

The calculation of the best IK can now be enabled/disabled. Comes in
handy when we do not care that much, or for example just want a
general IK check. Then we can skip the search for best IK, saving
computation.