colli: remove dependency on PCL and Eigen for removing old laser data
We have 2 laser points and even 1 reference point (the robot position),
so basically 2 vectors. Calculations are simple trigonometric, no need for
line-to-point projections and distance calculations (which are more costly
btw) and hence no need to be depending on PCL for this simple calculation.

The previous comparisons of opposing beams was errornous anyway, because
if for example the 2 points were less than 1-degree apart, but in different
quadrants (of the robot's local coordinate system), they would've been
considered to be beams in different directions.