jaco: goto_thread: only pop first target from queue if fully executed
This allows other threads that have access to the queue check that
as well, for example to decide if the arms are final or not
(which includes checking if the queue is empty or not). Otherwise the
goto_thread could decide that it is not final, but another thread would
not see this and could think the queue is empty and hence fully procesed.