eclipse_externals: reference interfaces by UID, API change!
Until now interfaces were referenced by ID only. This causes a problem
if you have linked pairs of interfaces which share the same ID (to match
them) but are of different types. Therefore, change the API of all
functions to require passing of an interface UID of the form Type::ID
instead. Since this does not change the arity of most functions,
problems will appear only at run-time. Therefore, existing code should
be adapted now. We have changed the internal housekeeping data structure
from a vector to a map allowing for faster access with less code.

Additionally, we have modified the semantics of bb_read/write_interface.
Instead of getting or setting a single value on a specific interface,
they now trigger reading of a specific interface. The new functions
bb_get and bb_set replace their functionality.

Along the way we have condensated the code to wasteful with white space
and newlines to make the code easier to read.