gui_utils: service chooser dialog remembers last manual host entry
Whenever the service chooser dialog was used it would have the manual
host entry expander closed and reverted to "localhost" as the default
entry text.

Now (if GConfmm is available) we remember the last expander state and
manual host entry. If the dialog is closed with the expander opened and
a non-empty host string it will be saved and next time the dialog is
opened (in any application) the expander will be opened with that text.
That makes operating GUIs a lot faster when working on a network or host
that does not allow for service discovery.