core: fixed waker waiting at barrier forever
When the 'waker' thread sets __wait_at_barrier to true and waits at the
Barrier, another thread can have a local_timeout and is waiting for the
mutex before it is woken up by the waker. Then the 'local_timeout' thread
could set __wait_at_barrier to false. Then the waker is the only thread
waiting at the Barrier.
I fixed this by letting only the waker set __wait_at_barrier to true.
It should not be necessary for the 'local_timeout' thread to clear
__wait_at_barrier because it is cleared when the interruptable barrier
is renewed.

Please test and review this commit before merging it.